+++ wls-gui-wahllokalsystem/src/views/wahlhandlung/BWBWahlbriefZulassungView.vue
@@ -37,8 +37,8 @@ import { storeToRefs } from "pinia";
 import { ref } from "vue";
 import { VTab, VTabs, VTabsWindow, VTabsWindowItem } from "vuetify/components";
 
+import TheBeanstandeteWahlbriefeBeschlussergebnis from "@/components/wahlhandlung/beanstandeteWahlbriefe/TheBeanstandeteWahlbriefeBeschlussergebnis.vue";
 import TheBeanstandeteWahlbriefeErfassenCard from "@/components/wahlhandlung/beanstandeteWahlbriefe/TheBeanstandeteWahlbriefeErfassenCard.vue";
-import TheBeanstandeteWahlbriefeBeschlussergebnis from "@/components/wahlvorbereitung/beanstandeteWahlbriefe/TheBeanstandeteWahlbriefeBeschlussergebnis.vue";
 import { useWahlenStore } from "@/stores/wahlenStore.ts";
 
 const { isBeanstandeteWahlbriefeTableValid } = storeToRefs(useWahlenStore());
